# Conflicts:
#	app/src/main/java/com/code4piter/blueskythinking/megapp/ui/activity/CamerasActivity.java
#	app/src/main/java/com/code4piter/blueskythinking/megapp/ui/activity/FilterActivity.java

+++ app/src/main/java/com/code4piter/blueskythinking/megapp/model/dto/RequestCameraListDto.java
@@ -17,9 +17,9 @@ public class RequestCameraListDto {
 
     private Double DangerLevel;
 
-	private Long latitude;
+	private Double latitude;
 
-	private Long longitude;
+	private Double longitude;
 
 	private String sortBy;
 
@@ -47,19 +47,19 @@ public class RequestCameraListDto {
 		DangerLevel = dangerLevel;
 	}
 
-	public Long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public Long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
